requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
Jinja2==3.1.2
pytube==15.0.0
PyYAML==6.0.1
requests==2.31.0
selenium==4.13.0
urlextract==1.8.0
